refactor: split TODO components and fix duplicate notifications
Break down large Todo.jsx into smaller, manageable components:
- TodoInput: Handle todo input with cute gradient styling
- IncompleteTodos: Display pending todos with random emojis
- CompletedTodos: Show completed todos with strikethrough styling
- styles.js: Centralized cute styling definitions
- utils.js: Shared utility functions for emojis and colors

Fix duplicate notification issue by removing onSearch handler
from Input.Search component, now only uses Form onFinish.
